Add support for /etc/timezone. Package-Manager: portage-2.2_pre5

+++ sys-libs/timezone-data/timezone-data-2008a.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2008 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-libs/timezone-data/timezone-data-2008a.ebuild,v 1.1 2008/03/10 15:23:32 vapier Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-libs/timezone-data/timezone-data-2008a.ebuild,v 1.2 2008/03/22 23:49:37 vapier Exp $
 
 inherit eutils toolchain-funcs flag-o-matic
 
@@ -52,21 +52,29 @@ src_install() {
 
 pkg_config() {
 	# make sure the /etc/localtime file does not get stale #127899
-	local tz=$(unset TIMEZONE ; source "${ROOT}"/etc/conf.d/clock ; echo ${TIMEZONE-FOOKABLOIE})
-	[[ -z ${tz} ]] && return 0
-	if [[ ${tz} == "FOOKABLOIE" ]] ; then
-		elog "You do not have TIMEZONE set in /etc/conf.d/clock."
-		if [[ ! -e ${ROOT}/etc/localtime ]] ; then
-			cp -f "${ROOT}"/usr/share/zoneinfo/Factory "${ROOT}"/etc/localtime
-			elog "Setting /etc/localtime to Factory."
-		else
-			elog "Skipping auto-update of /etc/localtime."
+	local tz src
+
+	if [[ -e ${ROOT}/etc/timezone ]] ; then
+		src="/etc/timezone"
+		tz=$(<"${ROOT}"/etc/timezone)
+	else
+		src="/etc/conf.d/clock"
+		tz=$(unset TIMEZONE ; source "${ROOT}"/etc/conf.d/clock ; echo ${TIMEZONE-FOOKABLOIE})
+		[[ -z ${tz} ]] && return 0
+		if [[ ${tz} == "FOOKABLOIE" ]] ; then
+			elog "You do not have TIMEZONE set in /etc/conf.d/clock."
+			if [[ ! -e ${ROOT}/etc/localtime ]] ; then
+				cp -f "${ROOT}"/usr/share/zoneinfo/Factory "${ROOT}"/etc/localtime
+				elog "Setting /etc/localtime to Factory."
+			else
+				elog "Skipping auto-update of /etc/localtime."
+			fi
+			return 0
 		fi
-		return 0
 	fi
 
 	if [[ ! -e ${ROOT}/usr/share/zoneinfo/${tz} ]] ; then
-		elog "You have an invalid TIMEZONE setting in /etc/conf.d/clock."
+		elog "You have an invalid TIMEZONE setting in ${src}"
 		elog "Your /etc/localtime has been reset to Factory; enjoy!"
 		tz="Factory"
 	fi
